move option_map_or_none to its own module
